#4 地震情報表示機能に対応
